requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
kombu==5.5.2
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
